+++ tripod/std/dust.f90
@@ -153,7 +153,7 @@ subroutine fi_diff(D, SigmaD, SigmaG, St, u, r, ri, Fi, Nr, Nm_s)
 end subroutine fi_diff
 
 
-subroutine vrel_brownian_motion(cs, m, T, vrel, Nr, Nm)
+subroutine vrel_brownian_motion(cs, m, T,mu, vrel, Nr, Nm)
     ! Subroutine calculates the relative particle velocities due to Brownian motion.
     ! Its maximum value is the sound speed.
     !
@@ -162,6 +162,7 @@ subroutine vrel_brownian_motion(cs, m, T, vrel, Nr, Nm)
     ! cs(Nr) : Sound speed
     ! m(Nr, Nm) : Particle masses
     ! T(Nr) : Temperature
+    ! mu(Nr) : mean molecular weight
     ! Nr : Number of radial grid cells
     ! Nm : Number of mass bins
     !
@@ -176,6 +177,7 @@ subroutine vrel_brownian_motion(cs, m, T, vrel, Nr, Nm)
     double precision, intent(in) :: cs(Nr)
     double precision, intent(in) :: m(Nr, Nm)
     double precision, intent(in) :: T(Nr)
+    double precision, intent(in) :: mu(Nr)
     double precision, intent(out) :: vrel(Nr, Nm, Nm)
     integer, intent(in) :: Nr
     integer, intent(in) :: Nm
@@ -194,7 +196,8 @@ subroutine vrel_brownian_motion(cs, m, T, vrel, Nr, Nm)
     do i = 1, Nm
         do j = 1, i
             do ir = 1, Nr
-                dum = min(sqrt(fac2(ir) * (m(ir, j) + m(ir, i)) / (m(ir, j) * m(ir, i))), cs(ir))
+                !dum = min(sqrt(fac2(ir) * (m(ir, j) + m(ir, i)) / (m(ir, j) * m(ir, i))), cs(ir))
+                dum = (16. * cs(ir)**2 * mu(ir)/(pi*m(ir,i)))**0.5d0
                 vrel(ir, j, i) = dum
                 vrel(ir, i, j) = dum
             end do
@@ -202,6 +205,244 @@ subroutine vrel_brownian_motion(cs, m, T, vrel, Nr, Nm)
     end do
 
 end subroutine vrel_brownian_motion
+
+
+subroutine vrel_radial_drift(vdriftmax, St,vrel, Nr, Nm)
+  ! Subroutine calculates the relative particle velocities due to radial drift.
+  !
+  ! Parameters
+  ! ----------
+  ! vdriftmax(Nr) : Maximum drift velocity
+  ! St(Nr, Nm) : Stokes number
+  ! Nr : Number of radial grid cells
+  ! Nm : Number of mass bins
+  !
+  ! Returns
+  ! -------
+  ! vrel(Nr, Nm) : Relative velocities
+
+  implicit none
+
+  double precision, intent(in)  :: vdriftmax(Nr)
+  double precision, intent(in)  :: St(Nr,Nm)
+  double precision, intent(out) :: vrel(Nr, Nm, Nm)
+  integer,          intent(in)  :: Nr
+  integer,          intent(in)  :: Nm
+  
+  integer :: ir, i, j
+  double precision :: dum
+
+  do i=1, Nm
+    do j=1, i
+      do ir=1, Nr
+        dum = abs(2d0 * vdriftmax(ir) * (St(ir,i)/(St(ir,i)**2 +1d0) -  St(ir,j)/(St(ir,j)**2 +1d0)))
+        vrel(ir, i, j) = dum
+        vrel(ir, j, i) = dum
+      end do
+    end do
+  end do
+
+end subroutine vrel_radial_drift
+
+
+subroutine vrel_vertical_settling(h, OmK, St, vrel, Nr, Nm)
+  ! Subroutine calculates the relative particle velocities due to vertical settling.
+  !
+  ! Parameters
+  ! ----------
+  ! h(Nr, Nm) : Particle scale heights
+  ! Omk(Nr) : Keplerian frequency
+  ! St(Nr, Nm) : Stokes number
+  ! Nr : Number of radial grid cells
+  ! Nm : Number of mass bins
+  !
+  ! Returns
+  ! -------
+  ! vrel(Nr, Nm) : Relative velocities
+
+  implicit none
+
+  double precision, intent(in)  :: h(Nr, Nm)
+  double precision, intent(in)  :: OmK(Nr)
+  double precision, intent(in)  :: St(Nr, Nm)
+  double precision, intent(out) :: vrel(Nr, Nm, Nm)
+  integer,          intent(in)  :: Nr
+  integer,          intent(in)  :: Nm
+
+  integer :: ir, i, j
+
+  double precision :: dum
+  double precision :: OmMinSt(Nr, Nm)
+
+  do i=1, Nm
+    do ir=1, Nr
+      OmMinSt(ir, i) = OmK(ir) * h(ir, i) * St(ir, i)/(St(ir, i) + 1d0)
+    end do
+  end do
+
+  do i=1, Nm
+    do j=1, i
+      do ir=1, Nr
+        dum = abs(OmMinSt(ir, j) - OmMinSt(ir, i))
+        vrel(ir, j, i) = dum
+        vrel(ir, i, j) = dum
+      end do
+    end do
+  end do
+
+end subroutine vrel_vertical_settling
+
+
+subroutine vrel_ormel_cuzzi_2007(alpha, cs, mump, OmegaK, rhoGas, &
+  & St, vrel, Nr, Nm)
+  ! Subroutine calculates the relative particle velocities due to turbulent motion
+  ! accourding the prescription of Ormel & Cuzzi (2007).
+  !
+  ! Parameters
+  ! ----------
+  ! alpha(Nr) : Turbulent alpha parameters
+  ! cs(Nr) : Sound speed
+  ! mump(Nr) : Mean molecular weight of the gas
+  ! OmegaK(Nr) : Keplerian frequency
+  ! rhoGas(Nr) : Gas midplane density
+  ! St(Nr, Nm) : Stokes number
+  ! Nr : Number of radial grid cells
+  ! Nm : Number of mass bins
+  !
+  ! Returns
+  ! -------
+  ! vrel(Nr, Nm) : Relative velocities
+
+  use constants, only: sigma_H2,pi
+
+  implicit none
+
+  double precision, intent(in)  :: alpha(Nr)
+  double precision, intent(in)  :: cs(Nr)
+  double precision, intent(in)  :: mump(Nr)
+  double precision, intent(in)  :: OmegaK(Nr)
+  double precision, intent(in)  :: rhoGas(Nr)
+  double precision, intent(in)  :: St(Nr, Nm)
+  double precision, intent(out) :: vRel(Nr, Nm, Nm)
+  integer,          intent(in)  :: Nr
+  integer,          intent(in)  :: Nm
+
+  double precision :: eps(Nr, Nm, Nm)
+  double precision :: OmKinv(Nr)
+  double precision :: Re
+  double precision :: ReInvSqrt(Nr)
+  double precision :: StL(Nr, Nm, Nm)
+  double precision :: StS(Nr, Nm, Nm)
+  double precision :: tauL(Nr, Nm, Nm)
+  double precision :: tauS(Nr, Nm, Nm)
+  double precision :: ts(Nr)
+  double precision :: vg2(Nr)
+  double precision :: vn
+  double precision :: vs(Nr)
+
+  double precision :: c0, c1, c2, c3, ya, yap1inv
+  double precision :: h1(Nr, Nm, Nm)
+  double precision :: h2(Nr, Nm, Nm)
+  double precision :: ys(Nr, Nm, Nm)
+
+  integer :: ir, i, j
+  double precision :: dum
+
+  c0      =  1.6015125d0
+  c1      = -0.63119577d0
+  c2      =  0.32938936d0
+  c3      = -0.29847604d0
+  ya      =  1.6d0
+  yap1inv =  1.d0/ (1.d0 + ya)
+
+  do ir=1, Nr
+    OmKinv(ir)    = 1.d0 / OmegaK(ir)
+    Re            =  alpha(ir) * rhoGas(ir) * sigma_H2 *cs(ir)/OmegaK(ir) / mump(ir)
+    ReInvSqrt(ir) = sqrt(1.d0 / Re)
+    vn            = sqrt(alpha(ir)) * cs(ir)
+    vs(ir)        = Re**(-0.25) * vn
+    ts(ir)        = OmKinv(ir) * ReInvSqrt(ir)
+    vg2(ir)       = 1.5d0 * vn**2
+  end do
+
+  do i=1, Nm
+    do j=1, i
+      do ir=1, Nr
+        StL(ir, j, i) = max(St(ir, j), St(ir, i))
+        StS(ir, j, i) = min(St(ir, j), St(ir, i))
+        eps(ir, j, i) = StS(ir, j, i) / StL(ir, j, i)
+
+        tauL(ir, j, i) = StL(ir, j, i) * OmKinv(ir)
+        tauS(ir, j, i) = StS(ir, j, i) * OmKinv(ir)
+
+        ys(ir, j, i) = c0 + c1*StL(ir, j, i) + c2*StL(ir, j, i)**2 &
+          & + c3*StL(ir, j, i)**3
+        h1(ir, j, i) = (StL(ir, j, i) - StS(ir, j, i))                 &
+          & / (StL(ir, j, i) + StS(ir, j, i))                          &
+          & * (StL(ir, j, i) * yap1inv                                 &
+          & - StS(ir, j, i)**2 / (StS(ir, j, i) + ya * StL(ir, j, i)))
+        h2(ir, j, i) = 2.d0 * (ya * StL(ir, j, i) - ReInvSqrt(ir))     &
+          & + StL(ir, j, i) * yap1inv                                     &
+          & - StL(ir, j, i)**2 / (StL(ir, j, i) + ReInvSqrt(ir))       &
+          & + StS(ir, j, i)**2 / (ya * StL(ir, j, i) + StS(ir, j, i))  &
+          & - StS(ir, j, i)**2 / (StS(ir, j, i) + ReInvSqrt(ir))
+      end do
+    end do
+  end do
+
+  do i=1, Nm
+    do j=1, i
+
+      where(tauL(:, j, i) .LT. 0.2d0 * ts(:))
+
+        vRel(:, j, i) = 1.5d0 * (vs(:) / ts(:) &
+          & * (tauL(:, j, i) - tauS(:, j, i)))**2
+
+      elsewhere(tauL(:, j, i)*ya .LT. ts(:))
+
+        vRel(:, j, i) = vg2(:) * (StL(:, j, i) - StS(:, j, i)) &
+          & / (StL(:, j, i) + StS(:, j, i)) * (StL(:, j, i)**2 &
+          & / (StL(:, j, i) + ReInvSqrt(:))                    &
+          & - StS(:, j, i)**2 / (StS(:, j, i) + ReInvSqrt(:)))
+
+      elsewhere(tauL(:, j, i) .LT. 5.d0 * ts(:))
+
+        vRel(:, j, i) = vg2(:) * (h1(:, j, i) + h2(:, j, i))
+
+      elsewhere(tauL(:, j, i) .LT. 0.2d0 * OmKinv(:))
+
+        vRel(:, j, i) = vg2(:) * StL(:, j, i)                             &
+          & * (2.d0*ya - 1.d0 - eps(:, j, i) + 2.d0/(1.d0 + eps(:, j, i)) &
+          & * (yap1inv + eps(:, j, i)**3/(ya+eps(:, j, i))))
+
+      elsewhere(tauL(:, j, i) .LT. OmKinv(:))
+        vRel(:, j, i) =  vg2(:) * StL(:, j, i)                  &
+          & * (2.d0*ys(:, j, i) - 1.d0 - eps(:, j, i)           &
+          & + 2.d0/(1.d0+eps(:, j, i))*(1.d0/(1.d0+ys(:, j, i)) &
+          & + eps(:, j, i)**3/(ys(:, j, i)+eps(:, j, i))))
+
+      elsewhere(tauL(:, j, i) .GE. OmKinv(:))
+
+        vRel(:, j, i) = vg2(:) &
+          & * (2.d0 + StL(:, j, i) + StS(:, j, i)) &
+          & / (1.d0 + StL(:, j, i) + StS(:, j, i) + StL(:, j, i)*StS(:, j, i))
+
+      end where
+
+    end do
+  end do
+
+  do i=1, Nm
+    do j=1, i
+      do ir=1, Nr
+        dum = sqrt(vRel(ir, j, i))
+        vRel(ir, j, i) = dum
+        vRel(ir, i, j) = dum
+      end do
+    end do
+  end do
+
+end subroutine vrel_ormel_cuzzi_2007
 
 
 subroutine calculate_m(a, rhos, fill, masses, Nr, Nm)
@@ -273,7 +514,7 @@ subroutine pfrag(vrel, vfrag, pf, Nr)
 end subroutine pfrag
 
 subroutine qfrag(p_dr, dv_tot, vfrag, St_max, q_turb1, q_turb2, &
-    & q_drfr, alpha, SigmaGas, mump, q_frag, Nr)
+    & q_drfr, alpha, rhoGas,cs,OmegaK, mump, q_frag, Nr)
     ! Subroutine calculates the power-law in the fragmentation
     ! regime, interpolating between different cases.
     !
@@ -291,7 +532,9 @@ subroutine qfrag(p_dr, dv_tot, vfrag, St_max, q_turb1, q_turb2, &
     ! q_turb1 : the power-law exponent for fragmentation in the first turbulence regime
     ! q_turb2 : same for the second turbulence regime
     ! alpha : the turbulence parameter
-    ! SigmaGas : the gas surface density
+    ! rhoGas : the gas surface density
+    ! cs : sound speed
+    ! OmegaK : angular speed
     ! mump : array of mean molecular mass (\mu * m_p)
     ! q_drfr :  same if drift is causing fragmentation
     ! Nr : Number or radial grid cells
@@ -309,7 +552,9 @@ subroutine qfrag(p_dr, dv_tot, vfrag, St_max, q_turb1, q_turb2, &
     double precision, intent(in) :: St_max(Nr)
     double precision, intent(in) :: q_turb1, q_turb2, q_drfr
     double precision, intent(in) :: alpha(Nr)
-    double precision, intent(in) :: SigmaGas(Nr)
+    double precision, intent(in) :: rhoGas(Nr)
+    double precision, intent(in) :: cs(Nr)
+    double precision, intent(in) :: OmegaK(Nr)
     double precision, intent(in) :: mump(Nr)
     double precision, intent(out) :: q_frag(Nr)
     integer, intent(in) :: Nr
@@ -320,7 +565,7 @@ subroutine qfrag(p_dr, dv_tot, vfrag, St_max, q_turb1, q_turb2, &
 
     do ir = 1, Nr
         !This seems wrong with the paper
-        Re = 1d0/sqrt(2d0*pi) * alpha(ir) * SigmaGas(ir) * sigma_H2 / mump(ir)
+        Re =  alpha(ir) * rhoGas(ir) *cs(ir)/OmegaK(ir) *  sigma_H2 / mump(ir)
 
         ! Eq. A.1 of Pfeil+2024
         f_t1t2 = 5d0 * sqrt(1.d0 / Re) / St_max(ir)
@@ -338,7 +583,7 @@ subroutine qfrag(p_dr, dv_tot, vfrag, St_max, q_turb1, q_turb2, &
 
 end subroutine qfrag
 
-subroutine pfrag_trans( St_max, alpha, SigmaGas, mump, p_frag_trans, Nr)
+subroutine pfrag_trans( St_max, alpha, rhoGas,cs,OmegaK, mump, p_frag_trans, Nr)
     ! Subroutine calculates the power-law in the fragmentation
     ! regime, interpolating between different cases.
     !
@@ -364,7 +609,9 @@ subroutine pfrag_trans( St_max, alpha, SigmaGas, mump, p_frag_trans, Nr)
 
     double precision, intent(in) :: St_max(Nr)
     double precision, intent(in) :: alpha(Nr)
-    double precision, intent(in) :: SigmaGas(Nr)
+    double precision, intent(in) :: rhoGas(Nr)
+    double precision, intent(in) :: cs(Nr)
+    double precision, intent(in) :: OmegaK(Nr)
     double precision, intent(in) :: mump(Nr)
     double precision, intent(out) :: p_frag_trans(Nr)
     integer, intent(in) :: Nr
@@ -375,7 +622,7 @@ subroutine pfrag_trans( St_max, alpha, SigmaGas, mump, p_frag_trans, Nr)
 
     do ir = 1, Nr
         !This seems wrong with the paper
-        Re = 1d0/sqrt(2d0*pi) * alpha(ir) * SigmaGas(ir) * sigma_H2 / mump(ir)
+        Re = alpha(ir) * rhoGas(ir) *cs(ir)/OmegaK(ir) *  sigma_H2 / mump(ir)
 
         ! Eq. A.1 of Pfeil+2024
         f_t1t2 = 5d0 * sqrt(1.d0 / Re) / St_max(ir)